Initial Website Upload
This upload represents the first step towards the ultimate goal of finishing this website and project. At present, email verification has been installed using SendGrid and work on Captcha implementation is ongoing. Tests have been excluded from this upload as they are in a separate project and are Selenium based; these tests are still in theoretical stage, as they have not been run against the project as of yet, due to a lack of a test-ready product.