Completion of Primary Scryfall Functionality
This update sees the completion of a scryfall search query builder as well as methods within the API with which to obtain one or more cards based on either CardID or a query. Some additional minor modifications also exist, related to issues found during testing.

Note: Not all tests that have been run may be present in this update.